Fix shared memory permission issue in a shared pod environment (#813)

+++ streaming/base/util.py
@@ -184,9 +184,12 @@ def clean_stale_shared_memory() -> None:
                 try:
                     shm = BuiltinSharedMemory(name, True, 4)
                 except FileExistsError:
-                    shm = BuiltinSharedMemory(name, False, 4)
-                    leaked_shm = True
-                finally:
+                    try:
+                        shm = BuiltinSharedMemory(name, False, 4)
+                        leaked_shm = True
+                    except PermissionError:
+                        continue
+                if shm:
                     shm.close()  # pyright: ignore
                     shm.unlink()
             # Come out of loop if no leaked shared memory
